bumped all relevant atari versions
