+++ pettingzoo/atari/othello_v1.py
@@ -1,0 +1,9 @@
+from .base_atari_env import BaseAtariEnv, base_env_wrapper_fn, parallel_wrapper_fn
+
+
+def raw_env(**kwargs):
+    return BaseAtariEnv(game="othello", num_players=2, mode_num=None, **kwargs)
+
+
+env = base_env_wrapper_fn(raw_env)
+parallel_env = parallel_wrapper_fn(env)
